Backen/requirements.txt
appdirs==1.4.3
Django==1.10.3
django-braces==1.10.0
django-cors-headers==1.3.1
django-oauth-toolkit==0.10.0
django-filter==1.0.4
djangorestframework==3.5.3
docutils==0.12
oauthlib==1.0.3
packaging==16.8
Pillow==3.4.2
PyMySQL==0.7.9
pyparsing==2.2.0
reportlab==3.4.0
six==1.10.0
